requirements.txt
# Python 3.13+
flask==3.1.3
mysql-connector-python==9.3.0
python-dotenv==1.2.2
werkzeug==3.1.3
google-genai==1.72.0
